feat(web): agent detail page
Refs: #182
