+++ apps/beeai-ui/src/modules/agents/components/AgentDetail.module.scss
