+++ apps/beeai-ui/src/modules/agents/detail/AgentDetailView.module.scss
@@ -14,27 +14,6 @@
  * limitations under the License.
  */
 
-import { Container } from '#components/layouts/Container.tsx';
-import { AgentDetailView } from '#modules/agents/detail/AgentDetailView.tsx';
-import { routes } from '#utils/router.ts';
-import { useNavigate, useParams } from 'react-router';
-
-type Params = {
-  agentName: string;
-};
-
-export function Agent() {
-  const { agentName } = useParams<Params>();
-  const navigate = useNavigate();
-
-  if (!agentName) {
-    navigate(routes.notFound(), { replace: true });
-    return null;
-  }
-
-  return (
-    <Container>
-      <AgentDetailView name={agentName} />
-    </Container>
-  );
+.tryButton {
+  min-inline-size: rem(208px);
 }
