+++ apps/beeai-web/src/@types/carbon__motion/index.d.ts
@@ -1,0 +1,38 @@
+/**
+ * Copyright 2025 IBM Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare module '@carbon/motion' {
+  export const fast01: string;
+  export const fast02: string;
+  export const moderate01: string;
+  export const moderate02: string;
+  export const slow01: string;
+  export const slow02: string;
+  // V11 Tokens
+  export const durationFast01: string;
+  export const durationFast02: string;
+  export const durationModerate01: string;
+  export const durationModerate02: string;
+  export const durationSlow01: string;
+  export const durationSlow02: string;
+
+  type EasingMode = 'productive' | 'expressive';
+  type EasingName = 'standard' | 'entrance' | 'exit';
+
+  export const easings: Record<EasingName, Record<EasingMode, string>>;
+
+  export function motion(name: EasingName, mode: EasingMode): string;
+}
