+++ apps/beeai-web/src/app/agents/page.tsx
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-import { getAcpClient } from "@/acp/client";
+import { getAgentsList } from "@/acp/api";
 import { Container, ViewStack } from "@i-am-bee/beeai-ui";
-import { unstable_cache } from "next/cache";
 import { AgentsView } from "./AgentsView";
 
 export default async function AgentsPage() {
@@ -29,18 +28,3 @@ export default async function AgentsPage() {
     </Container>
   );
 }
-
-const getAgentsList = unstable_cache(
-  async () => {
-    const client = await getAcpClient();
-
-    const { agents } = await client.listAgents();
-
-    await client.close();
-
-    return agents;
-  },
-  undefined,
-  // ISR can be added like this
-  // { revalidate: 30 }
-);
